backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cssselect==1.2.0
cssutils==2.7.1
distlib==0.3.6
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
filelock==3.12.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lxml==4.9.2
platformdirs==3.5.0
premailer==3.10.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
rsa==4.9
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.0
yagmail==0.15.293
